Move MainQueueDispatchDecorator to its own file
