+++ EssentialApp/EssentialApp/Comment/CommentUIComposer.swift
@@ -31,21 +31,3 @@ public final class CommentUIComposer {
 	}
 }
 
-class MainQueueDispatchDecorator: CommentLoader {
-	private let decoratee: CommentLoader
-	init(decoratee: CommentLoader) {
-		self.decoratee = decoratee
-	}
-	
-	func load(completion: @escaping (CommentLoader.Result) -> Void) -> CommentLoaderTask {
-		decoratee.load { result in
-			if Thread.isMainThread {
-				completion(result)
-			} else {
-				DispatchQueue.main.async {
-					completion(result)
-				}
-			}
-		}
-	}
-}
